requirements.txt
aws-cdk.core
aws-cdk.aws-s3
aws-cdk.aws-lambda
aws-cdk.aws-glue
aws-cdk.aws_s3_notifications
aws-cdk.aws-events
aws-cdk.aws-events-targets
pydantic
modules
pyyaml